todoapp/requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.1
django-environ==0.9.0
django-heroku==0.3.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.2
static3==0.7.0
whitenoise==6.2.0
